Adding the OnBoarding Screens
OnBoarding Screen is the screen that will appear when someone open up the app for the first time. This screen contains 3 pages that can be swipped left or right. These pages contains the gif images, onboarding titles and onboarding subtitles.
Also, The screen will have a static skip button that will take the user to the last page, static next button that will take the user to the next page of the current page, and a static Smooth Page Indicator that will act like a page navigator to show which page are we currently at.