Creating a Login Form
When on the last onboarding screen you hit next button then you will be navigated to the Login Form (All the previous pages saved in GetX will be removed using Get.offAll()).
The Login Screen will  contains 4 portions:
LoginHeader: This will contains the logo, loginTitle, loginSubTitle.
LoginForm: This will contains the TextFormField of email and password, a checkbox with Remember me Text, a forget password text, An Elevated Button of Login and And OutlineButton of Create Account
Form Divider: This portion contains a Divider that will divide the Form from the Footer along with some text in it.
Footer: In the footer, we have 2 IconButtons of google and facebook aligned in center.

Right Now, each and every Button is without any Functionality. We will add it later.

Also, we have created login_header and login_form files inside features -> screen.onboarding -> login -> widgets
And, we have created form_divider and social_buttons inside common -> widgets.login_signup because the same code will be used in creating signup form as well.