Creating a SignUp Screen
When on the Login Page, you hit the Create Account button then you will be redirected to the Sign Up Screen.
This Screen contains following things:
Signup Title
Form: This contains the first name, last name, username, email, phone, password, terms and condition checkbox and an elevated button named "Create Account".
Divider: The same divider as we used in Login Form
Footer: The same footer of Socials that we created in Login Form

Next Up, we are gonna make an email verification Screen and Success Screen