Verify Email Screen & Success Screen
After creating the signup screen, we need to create a verify email screen which should appear when the user hit create account button. After successfully verifying email, it should take us to the success screen to show that our email is verified & account is created successfully and take us back to Login Screen.

Both of the screen have image, some texts related to the screens and the elevated buttons to take us to the new screens.

Also, the verify email screen has the clear icon in its AppBar which takes us back to the create account screen.