Creating Forget Password Screen & Resend Password Screen
After creating the Email verification & Success Screen we will now create a Forget Password Screen.

In this Screen, we have a Texts, TextFormField to enter the email at which the reset password link will be send & finally we have the elevated button to submit the email.

After submitting the emial, we will be redirected to the reset password screen, where we will have a text showing that password has been sent and contains two buttons: One with the text done and other with the text resend email.

NOTE: We also have a cross icon at the appbar of both screens BUT the cross Icon at reset password screen will redirect us to login screen instead of the forget passord screen because there we have use the Get.off instead of Get.to.
Also, we have resue the same screen for reset password which we used for success screen but with one extra button in it.