Creating a Custom AppBar (A Container with two semi circles and curved Edges)
When a user login, then he is redirected to the home screen. This screen contains the navigation bar menu that we created earlier and now I created a container with curved edges.
This container is actually like a AppBar that a Scaffold have but this AppBar is custom made using container and I have not used the inbuild appBar property of Scaffold.

This custom made AppBar contains two things:

(1) A container that has two half circles at the right end of the container. For this purpose first I created a stack widget and then positioned two containers in it having radius equals to the size of the container so that i can get circle shape. Plus I have used a opacity to make background visible.

(2) The same container having two curved edges but in the opposite direction. To achieve that I uses the ClipPath widget to provide three clipPath. The one which is a curve starting from the left corner, a straight line after the curve ends and the 2nd curve that ends up to the end of right corner.

Next up, I will put some app designing in this custom AppBar.