Creating the Custom Appbar using PrefferedSizeWidget
After Creating the custom design for the AppBar, I added the required material for the AppBar.

First of all, as I want the desired height for my app bar therefore I use the separate class which implements the preferredSizeWidget property.
Uisng this, I created a custom AppBar having all the properties that a original AppBar holds like leadingIcon, actions, title Widget, show back arrow etc.

After that I inserts a Column having the title and subtitle of AppBar, and the Cart Icon at the end which contains the Counter as a Stacked element.
This Counter is like how many items we have put added the Cart.

Next up, I will create a SearchBar and a Horizontal ListView Builder