Creating a SearchBar and a Horizontal Scrollable List
In this commit, I have created a custom search bar using Container Widget and not the inbuilt SearchBar Widget. This is because i want to reuse the search bar at different locations.

Also, I have created a Horizontal Scrollable List View using ListView.builder. First I created a Column Widget and inside it I created a Title Text with a TextButton in a single Row. And then I created 6 lists of  Circular Image Icon using Container and listview builder and below each container having their respective category name.

Also, for the TextOverlay I have set the TextOverlay to elipsis so that I can get the dots If the text excles the size of container.