Created a Carousel Slider with dots
Up till now the header of the home page is completed and now i have started working on body of the home page.

In the body the first thing that i have created is  the Carousel Slider having 3 images as it's content.

For showing up that on which image we are currently on I have created a circular container having width 50 and height 6 and uses a for loop to detect how many contents are there in Carousel Slider.

After this, I used the GetxController to detect the change in the index of Carousel Slider and uses that index in the for loop to give background color  of blue to only that dot (Circular Container) whose index matches the loop index.