Creating a Product Card inside a GridView.builder
Right Below the banner with Carousel Slider, I created a Product Card having a Image on which discount Tag & favourite Icon is stacked.

And below that image is the detail of  the product like Title, brand, pricing etc.

To display multiple product card in the form of rows and columns I used the GridView.builder Widget.
This Widget helps me to display 2 cards in a row using cossAxisCount and Also Allowed me to give the card my own height using mainAxisExtent Property.