Creating a Complete Store Screen.
Previously I have finished the Home Screen. And now I have created a Store Screen.

This screen have a new functionality of having a nested scrollable thing. For this purpose I have used a NestedScrollView Widget. This widget has two things:
One is the content that will be pinned in the backgroud (SliverAppBar)
Other is the Body that will be shown below.

Whenever I scroll the below screen upward the body content gets scrolled over the pinned content.

Also the Body content has multiple tabs on it.

For that I uses the bottom property of SliverAppBar and uses the TabBar Widget inside it.
Also, for the tabs to work I have to wrap the scaffold with the DefaulytTabController

Now, to show the content of all the tabs in the body, I have used TabBarView Widget.

Contents inside the SLIVERAPPBAR and the BODY:

SLIVERAPPBAR:
This pinned content contains the Search Bar which we created in our previous commits, Section Heading having the view all action button and the display of 4 Feature Brands using Grid Layout

BODY:
The body contains the TabBarView and inside it there is a List View having two Brand Show Case Cards and 4 Product Vertical Card using Grid Layout.

Hence the Store Screen is completed.