Creating a Wishlist Screen
After creating the store screen, the next screen that I have created is the Wishlist Screen.

This screen is the most simplest screen and uses the previously used TGridLayour and the TProductVerticalCard that we uses on the Home Screen

Whenever the user click on the heart icon then that product must be added to the wishlist screen. This functionality will be done later when I will deal with backend.
Also, there is an add icon on the App Bar which will redirect us to the Home Screen.
Next up we are gonna create an Account Setting Scrreen.