Creating the Account Setting Screen
This screen takes the header same as the home screen which is having a blue background and clipped from below.

This contains a custom AppBar and after that a List  Tile having a leading image, title, subTitle and trailing icon.

After the header I used multiple List Tiles. Some tiles also have a Switch Widget as a trailing property.