Creating the profile screen
Whenever the user press the edit  button in front of profile pic inside the settings screen, the user is redirected to the Profile Screen.

This screen has the custom AppBar with only Profile Text.

The body has the Profile pic with textButton having change profile pic text.

After that a divider is used and below that is the Row having two expanded text with some flex and an icon which are wrapperd inside the Gesture Detector to have an onPressed property.

In the end, It also has the close account textButton.