Creating the image slider of Product Details Screen
Whenever someone click on the product from the home page, then they will be redirected to the Product Detail screen.

Initially I have created the  Custom Container having the AppBar with Favourite Icon, a large size image at the ceter of the contaner and the image slider at the bottom of the container.

Below this container, I have created Rating and share button, the discount badge with the discounted plus original value, title, stock status & the brand name with the icon.

Next up we are gonnna use the attributes value to provide different color scheme of the product, Size, checkout button etc