Creating a Rating & Review Screen
After completing the Product Detail Screen. Now I have created a Rating and Review screen which opens up when the user click on the icon next to review portion inside the product screen.

In this screen I have used many new things. First of all as i want to show the rating of the product therefore I have used the flutter_rating_bar package. In this package I have used the (1) RatingBarIndicator widget which is used to show the rating in the form of stars, (2) RatingProgressIndicator widget which is used to show the progress of rating in the forms of bar.

Also, there is the review portion in this screen where the user reviews about the product is shown. As these reviews can be longer than 3 lines therefore I have used another package named readmore.
Form this package, I have used the ReadMoreText widget which is sued to define the text, how many lines to trim the text after, morelabe, lesslabel, there styles etc.

Next-up, I am gonna create the addresses screen.