Creating the address screen
When the user goto the settings and click on the My Addresses then he will be redirected to the addresses screen.

This screen contains the conatiner having the name, phone number, address and the tick icon at the end of the container using the stack. This icon gets visible only if the particular address is selected.

Also this screen has the floating action button with the add icon. When the user click on it then he will be redirected to the new address screen where the user can add the new address to its account.

This new address screen has multiple form fileds like name, phone number, street, postal code etc and a save button.

Next-Up, I am gonna create all the screens related to settings and next I am gonna create the shopping Cart Screen