Creating the Cart Screen
Whenever the user clicks on the cart icon  on home screen then he will be redirected to the cart screen.

In this screen, I have placed list of items that are putted in the cart when he will clicks on the add button  on product card.

In this list each item will display the rounded image at the left after that in the column we will have the brand title with verified tick, Product name, color, size and in the last I have created two circular icon with add, minus icon with text in between.

At the end of the page, as a bottom navigation bar I have created an elevated button with a Checkout text along with the total price.

Next up I am gonna create the checkout screen whenever the user clicks on the check out button.