Creating the Checkout Screen
Whenever the user hits the checkout button on the CartScreen then he will be redirected to the CheckOut Screen.

In this screen I have created the overview of the items that a user have added to the cart but without the add and minus button.

Also, below that I  have created the textfield that can accept the coupon code.

Below that I have created the section where the details of pricing is provided, a divider, Payment method & the shipping address is provided.

At the end the checkout button is placed as a bottom navigation bar.

When the user hits this button then the success screen is displayed having the payment success text. This is the screen that we have already created before while creating the login pages.

At the end when the user hits continue button on the success screen then he is redirected to the home screen or we can say the navigation menu screen.

Next, I am going to create the order screen which is opened from the settings screen