Creating a Order Screen
In the settings screen, when the user clicks on my order then he is redirected to the order screen where he can see all the orders he has placed.

This screen is the simplest screen having the list of TRoundedContainer widget.

This widget has the column widget which has three row widgets. Each row has the icon on the left, title and subtitle as another column. One row has the icon at the end as well.

Next Up I am gonna create the sub categories screen which is displayed when the user clicks on any category displayed on home screen.