Creating the sub categories screen
This screen is displayed when the user clicks on any category that are shown in the home screen.

This screen contains the banner at the top. Also, down below there is a Horizontal Card. Previously on the home screen we have used the vertical product card but here we will use horizontal product card with few changes related to  the horizontal thing.

Also, the list are shown in the horizontal direction and not in the vertical direction

Next Up, I am gonna create the view all screen for the Products.