Creating the View All Product Screen
Whenever the user clicks on the view all button on the home screen then he will be redirected to the View all product sreen.

This screen is also the simplest screen which only has one new things and that is drop down menu.

This screen has the drop down menu at the topbwith some items in it and below it is the grid layout that we have created before.

We will use the same widget and our screen is completed.

Next up I am gonna create the view all screen for the Featured Brands which is going to be our last screen for the frontend