Creating the View All Brands Screen
Whenever the user click on the view all buttons on the store screen then he will be redirected to the View All Brands Screen.

This is the last screen for the frontend feature of our application.

This is also the simplest screen which utilizes all the previously created widgets.

This screen has the section heading without the action button. After that the grid layout list of brands card is used having the main axis extent reduced accordingly.

Also, when the user clicks on any of the brand then he is redirected to the all products screen related to that brands.

On that screen there is only one change then the previously created all product screen and that is that This screen contains the brand card of that specific brand at the top.

Finally, Our frontend part of e-commerce application is complete.

Next up, I am gonna start creating the Backend of the application