Setting the onBoardingScreen to appear only Once along with Splash Screen to preserve until everything loads from Firebase
After successful connecting the firebase to our flutter project. I am going to make my splash screen to stay preserved until the things of firebase is initialized to our project.

For that I have first initialize the GetStorage() and then preserve the splash screen using the FlutterSplashScreen.preserve() function.

Now I have created a separte class of AuthenticationRepository.

This class will have a onReady function which is called automatically when the AuthenticationRepository class is called.
This function will remove the Splash Screen

Until now, the  splash screen is preserved until everything loads up

Now, I have to set the onBoarding screen to appears only once. When the user reaches to login screen then the onBoarding screen should not be visible to user in the further app opening.

For this process I have used the GetStorage() function. This function has a writeIfNull function which creates the variable with a key and its value only if it does not exist in the GetStorage.

Then, I have used that same variable and has change its value to false when the user reaches to loginscreen.

Now at the start the onBoarding screen will appear but when the user reaches to login screen then the variable changes its value to false and the onBoarding screen will not appear.

Next up I am gonna create Email Password Authentication