Sign Up Form Validation
After creating a one Time onBoarding Screens, the next task  is to  create a form validation for the Sign Up Form Fields.
For this purpose, I first created a separate Sign Up Controller which carries all the text Editing Controllers for all the input fields
Then, I used all the controllers in the Sign Up form with key value as globalkey, first name input having controller with value of text editing controller of its respective one and so on.
Then, in validator property, I used the separate validator text for each input field and for that I created a separate class named TValidator. This class contains do the Validator check on each field and carries the function related to each input field e.g: if first name field is empty then it checks if the field is empty and displays the error of first name is required. Also, for checking the email, I used the RegExp() function that carries the character that should be available inside the email input field like it should first contain some alphabets then @ sign then alphabet then dot(.) then some alphabet.

After creating validators text for each input field then I have created the toggle password action. For this purpose I created the Rx<bool> type of variable for the toggle password. This variable is then used in the obscure text property of input field. and the same goes for the privacy policy check.

But in the privacy policy check If the privacy policy is not checked then it will pop the warning snackbar. For each kind of snackbar I created a separate class named TFullScreenLoader.

Finally the form Validation is completed and next up I am gonna Register the new user to the Firebase Authentication and also gonna save the data of new user to the Firebase Firestore.