Registering and saving data to Firebase Authentication & Firebase Firestore respectively
After validating the Sign Up form, the next thing I have done is to first register the new user to the Firebase Authentication using the instance of FirebaseAuth and then save the data of the new user to the Firebase Firestore using the instance of FirebaseFirestore.

First of all, I have added the two new packages named "firebase_auth" & "cloud_firestore".

To register the user, I have used the method of FirebaseAuth named "createUserWithEmailAndPassword(String emial, String password)". Also, for data to be fully registered, I have used await so that the code should not move forward until the registration is completed.

To save the data of new user, I have used the method of FirebaseFirestore named "collection(name).doc(id).set(toJson())".

Now, Whenver the user enter the data into the signup and clicks on the create button then, First the form is validated than if the validation is correct than the new user is registered to firebase and data is saved to firebase firestore.

Next up, I am gonna send the email to the user to verify the email address.