Sending email verification and setting timer for auto redirect
After successfully registering and storing data to the Firebase, I have implemented the verify-email controller for automatically sending the email for verification to the user whenever the user goes to the verify screen.

Also, I have created a timer which repeats the callback function every 1 seconds .

This function checks weather the user has verified the email or not. If yes, then the user will be redirected to the home screen.

If not, then if the user closes the application and re open it then the user will automatically be redirected to the verify screen and the email will be send.

Next up, We are going to work on the sign in form.