Login Form Validation & Authentication.
After successfully creating email verification of newly created account, the next step is to make form validation of login form same as we made for sign up form.
After that, I have use the current user of FirebaseAuth instance inside a new class named LoginController and using that currentUser I have made the choice of weather to the user is registered or not and then decide weather to login or first go for verifying email if not. This decision is coded inside the Authentication Repository.
After that I have Also added the sign out method inside the Authentication Repository which is called on the logout button inside the setting.dart file. This is done so that the current user can be logged out if the user wanted.

Next Up, I am gonna make the user sign in using the google sign in method.