Google Sign In Method
After successfully implementing the login validation, authentication and successfully logged in, the next step is to implement the Google Sign In method.

In this method, whenever the user clicks on the google icon below then the user will be redirected to the screen where he can select the google account for sign in.
For this purpose I have create the variable of type GoogleSignInAccount and It has GoogleSignIn().signIn() method.

After successfully selecting the google account, the next step is to fetch the google authentication detail of that account.
For this purpose I have create a new variable of type GoogleSignInAuthentication and it has the .authentication method of previously created GoogleSignInAccount variable.

After that, I have created a user credential from the fetched authentication detail of the account.
For this purpose, I have used the GoogleAuthProvider.credentials() method and inside its argument I have provided the idToken and accessToken of the GoogleSignInAuthentication variable.

Finally I have return the .signInWithCredential method of the FirebaseAuth.

In the end, I have stored the userCredential data to the FireStore.

Next Up, I am gonna create the forget password and reset password functionalities.