Forget Password & Reset Password
After successfully creating the login authentication, the next step is to create the functions for Resetting the password that if someone forget it.

For this purpose, I have first created a separate Getx class named ForgetPassswordController which has different variables for form validation

When the form is validated then we will call the .sendPasswordResetEmail() of _auth and as an argument pass the email that the user entered in a Forget Password form.

Also on the next screen which is a reset password screen there is a button that says re send email. When this button is clicked then again the same sendPasswordResetPassword function is called.

Hence the Password Reset is completed.

Up-till now, I have completed all the Firebase operations related to sign up and sign in. Next up, I am gonna starts dealing with the content inside the application after login.