Retrieving User Details from FireStore
For retrieving data from the fire store, we first have to fetch the data from the FirebaseFirestore.instance.collection(...).doc(uid).get().
This function returns the documentSnapshot of the current User and using that document snapshot we will create a UserModel.
This UserModel is the same model that we used to save the data in the firestore.

Now, I used that UserModel data and create a separate Rx<UserModel> variable that has the data of the UserModel. Using that variable I have made the Full name at the home screen and the profile details in the profile screen to dynamic.

Next Up, I am gonna create an update function which will update the First and the Last name of the user.