Updating the first and last name of User
After successfully fetching the data of the user, the next thing I have done is to create a controller that will update the First Name and the Last Name of the current User

Therefore First I have create a new screen named ChangeName which has a form in it having two input fields First Name & Last Name and also a save button.

When the user click on the > icon in front of the name in the ProfileScreen, then the user is redirected to the ChangeName Screen.

When the user clicks on the save button then first the form is validated and then the data is updated using the  function that we have created inside the ChangeNameController which again calls the updateSingleField function of the UserRepository.

After successfully updating the data in the Fire Store I have manually updated the value of the UserController.user.value.first/lastname with the first and last name textEditingController of changeNameController beacsue I only want to fetch the user data from Firestore once to reduce the no of reads.

Next Up, I am gonna create the function to remove the account of the user.