Deleting the user account
To delete a user account permanently, I have first created a removeUserRecord function inside the UserRepository which delete the data of the user from the FireStore.
After that, I have created few functions inside the UserController.

The first function is deleteAccountWarningPopup which shows the warning popup and ask for confirmation.

The second function is the deleteUserAccount which first fetch the provider of the user and check if provider is googl or email/passowrd. If google then first it call the signinWithGoogle function of Auth Repo and then calls the deleteAccount function of Auth Repo. Else If the provider is email/password then It calls the reAuthLoginForm.
In reAuthLoginForm, the user enter the email and password and clicks on verify. This will trigger the reAuthenticateWithEmailAndPassword function of UserController which re Authenticate the user and then calls the deleteAccount function of Auth Repo.

Finally, the account is deleted successfully.

Next up, I am gonna work on how to upload photos to the Firebase gallery.