Uploading Image to Firebase Storage + Updating Profile Pic + Shimmer Effect
After understanding CRUD Operations of Firebase, the next thing is to upload the image to the Firebase Storage and use that  Image as a Profile Pic.

For this purpose, first I have added few dependencies that are required. These includes : image_picker, firebase_storage, shimmer.

Using the image_picker we will first create a function that will first pick the image from the gallery. This uses ImagePicker().pickImage(source: ) function for fetching the image from gallery and then saving it to a variable.

Now using this image and the custom path, I have created a separate function in the UserReppository named 'uploadImage'. This function create a reference to the path and image and uses that ref creates the Image UrL.

Now back to the old function. Using this Image url we will update the ProfilePicture key value of the Firebase Firestore and the user of the UserController.

Also, While this uploading will take place, I have added the shimmer effect Widget using the shimmer.fromColor() and passing the required properties with Container as a child.

Additionally, I have used the cachedNetworkImage instead of simple networkImage in the TCircularImage. This is done to that the app will fetch the image for the first time and then store it in the memory for next use. This will reduce the memory allocation.

NOTE: I have came to know about the compatibility of Firebase related packages. All the packages like firebase_core, firebase_auth, cloud_firestore, firebase_storage must be compatible with each other. Otheriwse different kind of casting errors can come.

Next up, I am gonna create a function that will fetch the categories from firebase only once.