Dealing with Categories and fetching their data from FireStore
From Now onwards we will see how we can fetch the data from FireStore to display them on our app.

First of all, I have fetched the categories from the firestore. For such purpose, I have created

(1) CategoryRepository having function like getAllCategories which will fetch all the categories and store them in a RxList type of variable plus some UploadDummyData functions

(2) CategoryController having functions like fetchCategories which calls allCategories functiion of CategoryRepo and from that it stores all the categories to allCategories RXList tyope variable and only featured Categories to the featuredCategories RXList type of variable.

(3) CategoryModel which is a model of Category data available in FireStore.

(4) TFirebaseStorageService which is a service that deals with the storage of Firebase Firestore.
Also, There are few functions that deals with uploading Images and its data to the FiresStore + a function that  gets Image data from assets.

Finally, I have used them in THomeCategories where I have replaced the item Count of ListView to categories length, in the store screen where i have changed the Tab content to categories.

Next Up, I am gonna fetch the image from the FireStore and use them in app.