Fetching Banners Images from FireStore + Redirect to relevant screen
After successfull fetching the categories from the firebase, the next thing I do is fetching the banners data from the FireStore.
For this purpose I have done all the same things as I have done in fetching Categories.

I have created a separate model, controller and repository for the banners and have created the functions related to fetching banners from firestore.

The one extra thing that I have done here is redirecting to the relevant screen for each banner. This screen information is availabkle in firestore about each banner inside the key named targetscreen. And by using the Get.toNamed(), I have redirected to the target Screen.

ALso, I have created a separate function to specify the GetPage() and inside it I have tell the name and the page to route to.

Finally I have created a list of pages with names same as I have used in the GetPage() and have called them in the getPages property of GetMaterialApp inside the app.dart