Few changes to remove a build error
The error i encountered whille runing my app was of about a kotlin version mismatch error. This error was caused when i accidentl updated the dependenvies version and one of the dependency was causing this version mismatch error.

Therefore i have made few changes in androd/build.gradle , app/build.gradle, gradle.properties etc.

But still the error persists and now i am tring deleting the android folder and recreate it to check if the erorr removes or not.