Product Details uploaded and fetched + error of categories fetching resolved
After the Categories are fetched from the Firebase, the next step is to fetch the Products from the Firebase.

For the categories error, the uploading error was due to final not being used with the categories list variable inside dummy.dart. Also, the error related to image was because I was using asset image instead of Network image.

After that I have used the same Methods for uploading the dummy data of Products which I used for Categories.

The new things in this was that for fetching the List of String, I used the List<String>.from(data[' ']).
Also, for fetching the List of model, I used the (data[' '] as List<dynamic>).map((e) => model.fromJson(e))
Also, for all the nullable variables, I have used the ?? operator and tell the default values.

Also, I have created a separate functions for fetching the range or prices and for calculating the percentage of sale price from original price.

Next up, I am gonna use the values from Product collection into the Product Details screen.