Fetching sortable products data
After completing the products details screen data, the next thing I am gonna do is to fetch the sortable products data when the user clicks on the view all button of the popular products.

In this I have used some new functions or widgets.

First of all, I have used the FutureBuilder method to build the widgets based on the future method that I have created and pass that future method in the future property of this FutureBuilder. The result will be thne saved in the snapshot.

Then I have created a separte function inside the allProductController to first check the sort option and depending on that sort the fecthed product data and dsipaly it.

Next Up, I am gonna upload and fetched the brands data and it's relevant screens.