Fetching the brands
After fetching popular products data in the view all button, the next step I have done is to fetch all the brands

The steps will remains the same for fetching brands as we do in fetching the products.

The only thing that is changed here is that I have to fetch the products related to specific brands because I have to display the products that are related to specific brands.

For this purpose I have created a separate function that uses the brand Id and matches the brand Id with the Products.Brand.Id and display those products on the BrandProducts screen.

Next Up I am gonna fetch all the brands related to category and their respective products.