Wishlist or Favorites Items
In order to add the data as a favorites, I have first created an instance of bucket of GetStorage. Here bucket means to create GetStorage variable for a specific user and if the data is fetched by that user then only the wishlist related to that user will be shown.

To create the bucket, we need the user Id which is fetched when the user logged in to the application.

Also, I have created a separate helper function for dealing with all the functions related to GetStorage and for a single instance I have created a LocalStorage variable.

For wishList, first when the user clicks on the heart icon, the id of that product will be fetched and that id is used to save data to the storage at specific key having list of Map<String, dynamic> and after that using the key I have fetched all the products at that key.

And using that products, display them at wishlist screen.

Next up I am gonna fetched all the sub categories