Completing My Address Screen
For this screen we needed a sub collection for the specific user, Because we want the address to be saved for the specific user and display the relevant address to that user.

For this purpose, I have created a sub Collection named Addresses inside the current User which has it's own model. Using this sub Collection, I have fetched the user addresses details and shown them at Address screen.

For adding new address, I have created TextEditingControllers and using their values I have saved the new address data to the Addresses Collection.

Next Up I am gonna complete the Cart Management including payment gateways etc.