Handling Cart Management
In order to deal with cart, We need to know the types of products we can save in the cart.

The products that I can save are:
(1) A single type product
(2) Variable type products

(1) A single type products:
                                         In this type of product whenever the user clicks on the + icon in the vertical product card then that product is saved in the cartItem variable of the CartController.
   This cartItem is a CartItemModel which has many variables like productId, variationId, image, tite etc.
Before adding the items to the cart the product is first converted to CartItem Model and then it is saved in the cartItem variable.
Also, all the relevant details like no of cart item, total price, product quantity etc.
Also, this cartItem is then stored to Loacal Storage with key 'cartItem' after converting the cartItem data to Json.

(2) Variable type product:
                                       In this type of prouct, when the user clicks on the + icon then the user is redirected to the ProductDetails screen. On that screen, the user can see the + and - icon in the bottom with the Add to cart Button.
If the user increase the count and then clicks on the Add to Cart Button without selecting any of the variation then he will not be allowed to add the item to cart.

Moreover, there is a icon of cart at tha appbar of the home screen. When the user clicks on that icon then he is redirected to the cart screen. This screen displays all the products (single + variable including each variation) along with their no of product quantity. Also, you can increase or decrease the product count from that screen as well.

Next Up, I am gonna complete the Order Management.