Finishing my application by completing checkout & order management
The last thing I have implemented in the application is completing checkout & Order Management.
For this, I have used all the cart items and using them I have created the checkout page which carries the method of payment and the billing address.
Moreover, the final payment amount is shown by applying all the shipping and tax fees.
First the user select the payment method like paypal, master card etc. After that the user selects the billing address. The user can add a new billing address If the user wants.
When the user enters all the data and clicks on the continue button then the payment will get successful and the order list is shown in the order section.
This order section carries the information about all the orders that the user has placed uptil now. One order carries the order id, name, delivery date, order status and order date.

Finally, the complete Flutter E-Commerce application using Getx state management and Firebase as backend has comes to an end.